MARS-76 Create `AttributeViewButton`
